a new hstream-common-server package
