+++ cabal.project
@@ -18,6 +18,7 @@ packages:
   common/api/hstream-api-hs.cabal
   common/hstream/hstream-common.cabal
   common/stats/hstream-common-stats.cabal
+  common/server/hstream-common-server.cabal
 
   hstream-store
   hstream-diffflow
