+++ hstream/app/server.hs
@@ -35,6 +35,7 @@ import           ZooKeeper                        (withResource,
 
 import           HStream.Base                     (setupFatalSignalHandler)
 import           HStream.Common.ConsistentHashing (HashRing, constructServerMap)
+import           HStream.Common.Server.HashRing   (updateHashRing)
 import           HStream.Common.Types             (getHStreamVersion)
 import           HStream.Exception
 import           HStream.Gossip                   (GossipContext (..),
@@ -313,16 +314,3 @@ showVersion = do
   API.HStreamVersion{..} <- getHStreamVersion
   putStrLn $ "version: " <> T.unpack hstreamVersionVersion
           <> " (" <> T.unpack hstreamVersionCommit <> ")"
-
--- However, reconstruct hashRing every time can be expensive
--- when we have a large number of nodes in the cluster.
-updateHashRing :: GossipContext -> TVar (Epoch, HashRing) -> IO ()
-updateHashRing gc hashRing = loop (0,[])
-  where
-    loop (epoch, list)=
-      loop =<< atomically
-        ( do (epoch', list') <- getMemberListWithEpochSTM gc
-             when (epoch == epoch' && list == list') retry
-             writeTVar hashRing (epoch', constructServerMap list')
-             return (epoch', list')
-        )
